// apps/web/src/types/dashboard.types.ts
export interface DashboardCategory {
  _id: string;
  name: string;
  icon?: string;
  color?: string;
}
// ...
export type DashboardTransactionType = "INCOME" | "EXPENSE";

export type DashboardPaymentMethod =
  "CASH" | "CARD" | "UPI" | "BANK_TRANSFER" | "WALLET" | "CHEQUE" | "OTHER";

export interface DashboardRecentTransaction {
  _id: string;
  title: string;
  amount: number;
  type: DashboardTransactionType;
  currency: string;
  paymentMethod?: DashboardPaymentMethod | null;
  transactionDate: string;

  categoryId?: string | DashboardCategory | null;
  subcategoryId?: string | DashboardCategory | null;
}